add ElementSubdomainAndBoundaryModifier class that sits between ElementSubdomainModifier and ThresholdElementSubdomainModifier
This new class has code associated with the moving boundary: the code is simply taken from the old ElementSubdomainModifier
